Backend/requirements.txt
aniso8601==9.0.1
astroid==2.9.3
boto3==1.20.46
botocore==1.23.46
certifi==2021.10.8
click==8.0.3
elasticsearch==7.16.3
Flask==2.0.2
Flask-Cors==3.0.10
Flask-JWT-Extended==4.3.1
Flask-RESTful==0.3.9
google-cloud-vision==2.7.0
greenlet==1.1.2
isort==5.10.1
itsdangerous==2.0.1
Jinja2==3.0.3
jmespath==0.10.0
lazy-object-proxy==1.7.1
MarkupSafe==2.0.1
mccabe==0.6.1
mysql==0.0.3
mysqlclient==2.1.0
platformdirs==2.4.1
py-bcrypt==0.4
PyJWT==2.3.0
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
s3transfer==0.5.0
six==1.16.0
SQLAlchemy==1.4.30
toml==0.10.2
typing_extensions==4.0.1
urllib3==1.26.8
Werkzeug==2.0.2
wrapt==1.13.3
